third_party/blink/web_tests/html/details_summary/details-click-controls-expected.txt

CONSOLE ERROR: An interactive element was found within a <summary> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology.
CONSOLE ERROR: An interactive element was found within a <summary> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology.
CONSOLE ERROR: An interactive element was found within a <summary> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology.
PASS details.open is false
PASS details.open is false
PASS details.open is false
PASS details.open is false
PASS details.open is true
PASS successfullyParsed is true

TEST COMPLETE
